Exception mapping
Wrap the Spark exception mapping to enable Kotlin syntax sugar.
