+++ .gitignore
@@ -2,6 +2,7 @@
 /build/
 target
 .idea
+*.iml
 
 # Ignore Gradle GUI config
 gradle-app.setting
