+++ src/main/kotlin/spark/kotlin/ExceptionHandler.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2017 Love Löfdahl
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package spark.kotlin
+
+import spark.Request
+import spark.Response
+
+/**
+ * Used to facilitate custom exception mapping.
+ *
+ * @param exception The exception that caused the handler to be triggered.
+ * @param request The Spark [Request]
+ * @param response The Spark [Response]
+ */
+class ExceptionHandler(val exception: Exception, request: Request, response: Response) : RouteHandler(request, response)
